Refactor timeline view into separate modules
