+++ ui/constants.py
@@ -1,0 +1,2 @@
+PIXELS_PER_SECOND = 100
+INITIAL_DURATION = 60  # seconds
